+++ src/cpp/desktop/DesktopRVersion.cpp
@@ -1,7 +1,7 @@
 /*
  * DesktopRVersion.cpp
  *
- * Copyright (C) 2009-18 by RStudio, Inc.
+ * Copyright (C) 2009-17 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -241,7 +241,7 @@ void enumRegistry(Architecture architecture, HKEY key, QList<RVersion>* pResults
    }
 
    std::vector<std::string> keys = regKey.keyNames();
-   for (size_t i = 0; i < keys.size(); i++)
+   for (int i = 0; i < keys.size(); i++)
    {
       RegistryKey verKey;
       error = verKey.open(regKey.handle(),
